backend/src/main.py
from dotenv import load_dotenv
load_dotenv()   # 🔥 MUST BE FIRST

from fastapi import FastAPI
from sqlalchemy import text
from src.infrastructure.db.base import Base
from src.infrastructure.db import models
from fastapi.middleware.cors import CORSMiddleware
from fastapi.staticfiles import StaticFiles
import os

# DB
from src.infrastructure.db.session import engine

# ROUTERS
from src.interfaces.api.v1.lead_router import router as lead_router
from src.interfaces.api.v1.voice_router import router as voice_router

# METRICS
from prometheus_client import make_asgi_app

app = FastAPI(
    title="Zeva Voice Demo Engine",
    version="1.0.0",
)

app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "http://localhost:3000",
    ],
    allow_credentials=True,
    allow_methods=["*"],   # IMPORTANT
    allow_headers=["*"],   # IMPORTANT
)

# -----------------------------
# PROMETHEUS METRICS
# -----------------------------
metrics_app = make_asgi_app()
app.mount("/metrics", metrics_app)


# -----------------------------
# ROUTES
# -----------------------------
app.include_router(lead_router, prefix="/api/v1", tags=["Leads"])
app.include_router(voice_router, prefix="/api/v1", tags=["Voice"])

AUDIO_DIR = "storage/audio"
os.makedirs(AUDIO_DIR, exist_ok=True)
app.mount("/audio", StaticFiles(directory="storage/audio"), name="audio")
# -----------------------------

@app.on_event("startup")
async def create_tables():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)
# -----------------------------
# HEALTH CHECK
# -----------------------------
@app.get("/health")
async def health():
    return {"status": "ok"}


@app.get("/health/db")
async def health_db():
    async with engine.begin() as conn:
        await conn.execute(text("SELECT 1"))
    return {"status": "db ok"}