backend/pyproject.toml
[project]
name = "zeva-backend"
version = "0.0.0"

dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy>=2.0",
    "asyncpg",
    "alembic",
    "psycopg2-binary",
    "pydantic",
    "python-dotenv",
    "redis",
    "pybreaker",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==3.2.2"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"